# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Todo Application REST-API/gRPC schema",
        "title": "Golang Todo API",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:5050",
    "basePath": "/api/v1",
    "paths": {
...
        "/users/{id}": {
            "get": {
                "description": "Получение пользователя по его ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Получить пользователя",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID получаемого пользователя",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "успешное получение пользователя",
                        "schema": {
                            "$ref": "#/definitions/users_transport_http.GetUserResponse"
                        }
                    },
                    "400": {
                        "description": "BadRequest",
                        "schema": {
                            "$ref": "#/definitions/core_http_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/core_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/core_http_response.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удалить пользователя из системе по его ID",
                "tags": [
                    "users"
                ],
                "summary": "Удалить пользователя",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID удаляемого пользователя",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "успешное удаление пользователя"
                    },
                    "400": {
                        "description": "BadRequest",
                        "schema": {
                            "$ref": "#/definitions/core_http_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/core_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/core_http_response.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Изменение информации об уже существующем пользователе\n### Логика обновления полей (Three-state logic):\n1.**Поле не передано**: `phone_number` игнорируется, значение в БД не меняется\n2.**Явно передано значение**: `\"phone_number\":\"+79876543210\"` - устанавливает номер телефона в бд\n3.**Передан null**: `\"phone_number\": null` - очищает поле в БД (set to NULL)\nОграничения: `full_name` не может быть выставлен как NULL",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Изменить задачу",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID изменяемого пользователя",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "PatchUser тело запроса",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/users_transport_http.PatchUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "успешно изменённый пользователь",
                        "schema": {
                            "$ref": "#/definitions/users_transport_http.PatchUserResponse"
                        }
                    },
                    "400": {
                        "description": "BadRequest",
                        "schema": {
                            "$ref": "#/definitions/core_http_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/core_http_response.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/core_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/core_http_response.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "core_http_response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "error text"
                },
                "message": {
                    "type": "string",
                    "example": "readable error message"
                }
            }
        },
...
        "users_transport_http.GetUserResponse": {
            "type": "object",
            "properties": {
                "full_name": {
                    "type": "string",
                    "example": "Alex Smith"
                },
                "id": {
                    "type": "integer",
                    "example": 11
                },
                "phone_number": {
                    "type": "string",
                    "example": "+79876543211"
                },
                "version": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "users_transport_http.PatchUserRequest": {
            "type": "object",
            "properties": {
                "full_name": {
                    "type": "string",
                    "example": "Max Snow"
                },
                "phone_number": {
                    "type": "string",
                    "example": "+79876543210"
                }
            }
        },
        "users_transport_http.PatchUserResponse": {
            "type": "object",
            "properties": {
                "full_name": {
                    "type": "string",
                    "example": "Alex Smith"
                },
                "id": {
                    "type": "integer",
                    "example": 11
                },
                "phone_number": {
                    "type": "string",
                    "example": "+79876543211"
                },
                "version": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
    }
}